# static/openapi/registry.yaml
openapi: 3.1.0
info:
  title: AgenticAdvertising.org Registry API
  description: |-
    Public REST API for the AgenticAdvertising.org registry. Resolve brands,
    discover properties, look up agents, and validate authorization in the
    AdCP ecosystem.

    All endpoints listed here are public and require no authentication.

    **Base URL:** `https://agenticadvertising.org`

    **Rate limits:** Bulk resolve endpoints are limited to 20 requests per minute
    per IP address. All other endpoints are unmetered.
  version: 1.0.0
  contact:
    name: AgenticAdvertising.org
    url: https://agenticadvertising.org
servers:
  - url: https://agenticadvertising.org
    description: Production
security: []
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
paths:
...
  /api/brands/enrich:
    get:
      operationId: enrichBrand
      summary: Enrich brand
      description: Enrich brand data using Brandfetch. Returns logo, colors, and company information.
      tags:
        - Brand Resolution
      parameters:
        - schema:
            type: string
            example: acmecorp.com
          required: true
          name: domain
          in: query
      responses:
        "200":
          description: Enrichment data from Brandfetch
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "503":
          description: Brandfetch not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...